+++ src/test/resources/application-test.yml
@@ -17,6 +17,13 @@ spring:
     username: sa
     password:
 
+    review:
+      raw:
+        driver-class-name: org.h2.Driver
+        url: jdbc:h2:mem:test
+        username: sa
+        password:
+
   jpa:
     hibernate:
       ddl-auto: create-drop
